Replaced updated pagination for all entities
