+++ src/app/layouts/admin-layout.tsx
@@ -16,18 +16,18 @@ import { Admin, Edit, Layout, Resource, Show } from 'react-admin'
 import { authProviderWithRefresh } from '@/app/providers/auth-provider'
 import { dataProviderWithRefresh } from '@/app/providers/data-provider'
 import { theme } from '@/app/ui/admin-theme'
-import { EditAdvantage } from '@/views/admin/advantages/edit.page'
-import { AdvantagesList } from '@/views/admin/advantages/list.page'
-import { EditCategory } from '@/views/admin/categories/edit.page'
-import { CategoriesList } from '@/views/admin/categories/list.page'
+import { EditAdvantage } from '@/views/admin/advantages/ui/edit.page'
+import { AdvantagesList } from '@/views/admin/advantages/ui/list.page'
+import { EditCategory } from '@/views/admin/categories/ui/edit.page'
+import { CategoriesList } from '@/views/admin/categories/ui/list.page'
 import { EditProduct } from '@/views/admin/dishes/ui/edit.page'
 import { ProductList } from '@/views/admin/dishes/ui/list.page'
-import { EditIngredient } from '@/views/admin/ingredients/edit.page'
-import { IngredientList } from '@/views/admin/ingredients/list.page'
-import { ShowOrder } from '@/views/admin/orders/id.page'
-import { OrdersList } from '@/views/admin/orders/list.page'
-import { EditPost } from '@/views/admin/posts/edit.page'
-import { PostsList } from '@/views/admin/posts/list.page'
+import { EditIngredient } from '@/views/admin/ingredients/ui/edit.page'
+import { IngredientList } from '@/views/admin/ingredients/ui/list.page'
+import { ShowOrder } from '@/views/admin/orders/ui/id.page'
+import { OrdersList } from '@/views/admin/orders/ui/list.page'
+import { EditPost } from '@/views/admin/posts/ui/edit.page'
+import { PostsList } from '@/views/admin/posts/ui/list.page'
 import { ChangePasswordModal } from '@/features/admin/change-password-modal'
 import { AdminAppBar } from '@/shared/ui/admin/admin-bar'
