+++ src/views/admin/advantages/ui/edit.page.tsx
