+++ src/views/admin/advantages/ui/index.ts
