+++ src/views/admin/categories/index.ts
@@ -1,0 +1,1 @@
+export * from './ui'
