+++ src/views/admin/categories/ui/edit.page.tsx
