+++ src/views/admin/categories/ui/list.page.tsx
@@ -1,10 +1,10 @@
-import { Dish } from '@/entities/dish'
+import { Category } from '@/entities/category'
 import { useMedia } from '@/shared/lib/hooks/useMedia'
 import { useResourcePaginationSortController } from '@/shared/lib/hooks/useResourcePaginationSort'
 import { DesktopList } from './desktop-list'
 import { MobileList } from './mobile-list'
 
-export const ProductList = () => {
+export const CategoriesList = () => {
   const {
     paginatedRows,
     currentPage,
@@ -12,7 +12,7 @@ export const ProductList = () => {
     setCurrentPage,
     limitParam,
     setLimit,
-  } = useResourcePaginationSortController<Dish>({ resource: 'dish' })
+  } = useResourcePaginationSortController<Category>({ resource: 'category' })
   const { isMobileScreen } = useMedia()
 
   return (
