+++ src/views/admin/ingredients/ui/edit.page.tsx
