+++ src/views/admin/ingredients/ui/index.ts
