+++ src/views/admin/orders/ui/index.ts
@@ -1,4 +1,5 @@
-export * from './edit.page'
+export * from './Empty'
+export * from './id.page'
 export * from './list.page'
 export * from './desktop-list'
 export * from './mobile-list'
