+++ src/views/admin/posts/ui/edit.page.tsx
